+++ MarvelComicsAlamofire/Resources/Assets.xcassets/Contents.json
