+++ MarvelComicsAlamofire/Sources/App/SceneDelegate.swift
@@ -1,0 +1,21 @@
+//
+//  SceneDelegate.swift
+//  MarvelComicsAlamofire
+//
+//  Created by Kristina Korotkova on 08/03/23.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        window = UIWindow(windowScene: windowScene)
+        let viewController = ComicViewController()
+        window?.rootViewController = viewController
+        window?.makeKeyAndVisible()
+    }
+}
